+++ packages/storage/src/singleton.rs
@@ -7,6 +7,9 @@ use crate::length_prefixed::to_length_prefixed;
 use crate::type_helpers::{may_deserialize, must_deserialize};
 
 /// An alias of Singleton::new for less verbose usage
+#[deprecated(
+    note = "The crate cosmwasm-storage is unmaintained and will be removed in CosmWasm 2.0. Please consider migrating to cw-storage-plus or simple cosmwasm-std storage calls."
+)]
 pub fn singleton<'a, T>(storage: &'a mut dyn Storage, key: &[u8]) -> Singleton<'a, T>
 where
     T: Serialize + DeserializeOwned,
@@ -15,6 +18,9 @@ where
 }
 
 /// An alias of ReadonlySingleton::new for less verbose usage
+#[deprecated(
+    note = "The crate cosmwasm-storage is unmaintained and will be removed in CosmWasm 2.0. Please consider migrating to cw-storage-plus or simple cosmwasm-std storage calls."
+)]
 pub fn singleton_read<'a, T>(storage: &'a dyn Storage, key: &[u8]) -> ReadonlySingleton<'a, T>
 where
     T: Serialize + DeserializeOwned,
@@ -26,6 +32,9 @@ where
 /// work on a single storage key. It performs the to_length_prefixed transformation
 /// on the given name to ensure no collisions, and then provides the standard
 /// TypedStorage accessors, without requiring a key (which is defined in the constructor)
+#[deprecated(
+    note = "The crate cosmwasm-storage is unmaintained and will be removed in CosmWasm 2.0. Please consider migrating to cw-storage-plus or simple cosmwasm-std storage calls."
+)]
 pub struct Singleton<'a, T>
 where
     T: Serialize + DeserializeOwned,
@@ -89,6 +98,9 @@ where
 
 /// ReadonlySingleton only requires a Storage and exposes only the
 /// methods of Singleton that don't modify state.
+#[deprecated(
+    note = "The crate cosmwasm-storage is unmaintained and will be removed in CosmWasm 2.0. Please consider migrating to cw-storage-plus or simple cosmwasm-std storage calls."
+)]
 pub struct ReadonlySingleton<'a, T>
 where
     T: Serialize + DeserializeOwned,